+++ packages/logger/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @graphql-yoga/logger
 
+## 2.0.1
+
+### Patch Changes
+
+- [#3660](https://github.com/dotansimha/graphql-yoga/pull/3660)
+  [`d4cbae1`](https://github.com/dotansimha/graphql-yoga/commit/d4cbae12348a39a2f64d7048dc582ebbaac93a5b)
+  Thanks [@renovate](https://github.com/apps/renovate)! - dependencies updates:
+  - Updated dependency [`tslib@^2.8.1` ↗︎](https://www.npmjs.com/package/tslib/v/2.8.1) (from
+    `^2.5.2`, in `dependencies`)
+
 ## 2.0.0
 
 ### Major Changes