+++ packages/nestjs-federation/CHANGELOG.md
@@ -1,5 +1,38 @@
 # @graphql-yoga/nestjs-federation
 
+## 3.10.11
+
+### Patch Changes
+
+- [#3676](https://github.com/dotansimha/graphql-yoga/pull/3676)
+  [`c9b167b`](https://github.com/dotansimha/graphql-yoga/commit/c9b167bc8fdf8478649f6507426b336836098c11)
+  Thanks [@renovate](https://github.com/apps/renovate)! - dependencies updates:
+  - Updated dependency
+    [`@nestjs/graphql@^13.0.0` ↗︎](https://www.npmjs.com/package/@nestjs/graphql/v/13.0.0) (from
+    `^12.0.0`, in `peerDependencies`)
+- Updated dependencies
+  [[`c9b167b`](https://github.com/dotansimha/graphql-yoga/commit/c9b167bc8fdf8478649f6507426b336836098c11)]:
+  - @graphql-yoga/nestjs@3.10.11
+
+## 3.10.10
+
+### Patch Changes
+
+- [#3666](https://github.com/dotansimha/graphql-yoga/pull/3666)
+  [`356d5f2`](https://github.com/dotansimha/graphql-yoga/commit/356d5f2be30b6a73fc1a4d7efab57a4213728684)
+  Thanks [@renovate](https://github.com/apps/renovate)! - dependencies updates:
+  - Updated dependency
+    [`@nestjs/common@^11.0.0` ↗︎](https://www.npmjs.com/package/@nestjs/common/v/11.0.0) (from
+    `^10.0.0`, in `peerDependencies`)
+  - Updated dependency
+    [`@nestjs/core@^11.0.0` ↗︎](https://www.npmjs.com/package/@nestjs/core/v/11.0.0) (from
+    `^10.0.0`, in `peerDependencies`)
+- Updated dependencies
+  [[`356d5f2`](https://github.com/dotansimha/graphql-yoga/commit/356d5f2be30b6a73fc1a4d7efab57a4213728684),
+  [`d4cbae1`](https://github.com/dotansimha/graphql-yoga/commit/d4cbae12348a39a2f64d7048dc582ebbaac93a5b)]:
+  - @graphql-yoga/nestjs@3.10.10
+  - @graphql-yoga/plugin-apollo-inline-trace@3.10.10
+
 ## 3.10.9
 
 ### Patch Changes